Web: Return the user id when creating an account Web: Implementation of GCM, to handle Android paste command Web: Initial implementation of MIME support Android: Handle the GCM registration on the server. Android: Push copied text to the server

+++ Web/Projecto/src/main/java/pt/andre/projecto/Controllers/IAPI.java
@@ -3,11 +3,11 @@ package pt.andre.projecto.Controllers;
 import org.springframework.http.ResponseEntity;
 
 public interface IAPI {
-    ResponseEntity push(Integer token, String data);
-    ResponseEntity pull(Integer token);
+    ResponseEntity push(long token, String data);
+    ResponseEntity pull(long token);
     ResponseEntity createAccount(String account, String password);
     ResponseEntity authenticate(String account, String password);
-    ResponseEntity registerAndroidDevice(Integer account, String deviceID);
+    ResponseEntity registerAndroidDevice(long account, String deviceID);
