+++ Web/Projecto/src/main/java/pt/andre/projecto/Service/Interfaces/IAPIService.java
@@ -5,13 +5,13 @@ import pt.andre.projecto.Model.Database.Utils.DatabaseResponse;
 
 public interface IAPIService {
 
-    DatabaseResponse push(int token, String data);
-    DatabaseResponse push(int token, String data, boolean isMIME);
+    DatabaseResponse push(long token, String data);
+    DatabaseResponse push(long token, String data, boolean isMIME);
 
     //Only used by the android client notification
-    DatabaseResponse registerAndroidDevice(int token, String firebaseID);
+    DatabaseResponse registerAndroidDevice(long token, String firebaseID);
 
-    DatabaseResponse pull(int token);
+    DatabaseResponse pull(long token);
     DatabaseResponse createAccount(String account, String password);
     DatabaseResponse authenticate(String account, String password);
